:construction: Basic toggle switch plus entry card selection.
currently cards are the same, but now I have room to edit.
